WIP: LNS displaying owned records
